Add files via upload
This file uses Java Swing.

This is the scaffolding of the checkers game I've been working on recently. I'm only an hour or so into the project, but so far I've been able to print an array of buttons that return their position on a grid system. The next step is to add more variables/characteristics to the BoardSpace class, so I can have pieces move around and jump over each other. There are lots of scatterbrained comments left over in the program, which I will make sure to remove in the final version! They are mostly to help me keep track of what I want to do next.